web/djangoAngular/requirements.txt
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
Markdown==2.6.9
psycopg2==2.7.3.1
pytz==2017.2
redis==2.10.3
django-redis==4.3.0
gunicorn==19.3.0
django-cors-headers